test_web/legion_bot/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2017.11.5
certificator==0.0.5
cffi==1.11.4
chardet==3.0.4
click==6.7
cssselect2==0.2.1
cx-Oracle==5.3
DBUtils==1.2
defusedxml==0.5.0
dill==0.2.7.1
html5lib==1.0.1
httpcache==0.1.3
idna==2.6
Jinja2==2.10
log==0.0.2
logger==1.4
lxml==4.1.1
MarkupSafe==1.0
MultipartPostHandler==0.1.0
multiprocessframe==0.0.1
numpy==1.14.0
pandas==0.22.0
pdfrw==0.4
Pillow==5.0.0
prettyconf==1.2.3
PyAutoIt==0.4
pycparser==2.18
pyHook==1.5.1
PyMouse==1.0
PyMySQL==0.8.0
Pyphen==0.9.4
pypiwin32==220
pyreadline==2.1
pysimplemodel==1.0.2
python-dateutil==2.6.1
python-xlib==0.21
pytils==0.3
pytz==2017.3
PyUserInput==0.1.11
pywin32==221
reppy==0.4.9
requests==2.18.4
requests-middleware==0.1.2
selenium==3.8.1
selenium-aurelia==0.1.1
six==1.11.0
sixfour==1.3.3
SQLAlchemy==1.2.2
tinycss2==0.6.1
tools==0.1.9
urllib3==1.22
WeasyPrint==0.42
webencodings==0.5.1
xlrd==1.1.0
